chore: estabilização da integração front-back e validação de contratos
- Core: Correção das discrepâncias de contrato D1 (Decimal -> float) e D2 (tipagem de métricas).
- Infra: Criação de Dockerfiles e ajuste do docker-compose para build automatizado.
- DB: Modelos SQLAlchemy migrados para tipos agnósticos (UUID/JSON) para suporte a SQLite nos testes.
- QA Front: Implementação de suíte de testes com Vitest + MSW (4/4 testes passando).
- QA Back: Implementação de testes de integração via Pytest + Docker (32/32 testes passando).
- Auth: Simplificação temporária da verificação de senha para garantir estabilidade no container.

+++ src/back/app/schemas/analysis.py
@@ -25,11 +25,11 @@ class TrechoChave(BaseModel):
 
 
 class PropostaAcordoResponse(BaseModel):
-    valor_sugerido: Decimal
-    intervalo_min: Decimal
-    intervalo_max: Decimal
-    custo_estimado_litigar: Decimal
-    economia_esperada: Decimal
+    valor_sugerido: float
+    intervalo_min: float
+    intervalo_max: float
+    custo_estimado_litigar: float
+    economia_esperada: float
     n_casos_similares: int
 
 
@@ -48,5 +48,5 @@ class AnaliseIAResponse(BaseModel):
 
 class DecisaoAdvogadoRequest(BaseModel):
     acao: AcaoAdvogado
-    valor_advogado: Decimal | None = None
+    valor_advogado: float | None = None
     justificativa: str | None = None
